+++ tests/testthat/helpers.R
@@ -1,0 +1,8 @@
+if (!dir.exists("travis-testthat")) {
+  system("git clone https://github.com/pat-s/travis-testthat.git")
+}
+
+# change to "travis-testthat"
+setwd("./travis-testthat")
+
+repo <- "travis-testthat"